chore: move cta component to common
