+++ src/components/pages/state/long-term-care/overview.js
@@ -3,10 +3,7 @@ import { useStaticQuery, graphql } from 'gatsby'
 import { Row, Col } from '~components/common/grid'
 import Total from '~components/common/landing-page/total'
 import { DefinitionPanel } from '~components/pages/data/cards/definitions-panel'
-import {
-  CtaLink,
-  CtaAnchorLink,
-} from '~components/common/landing-page/call-to-action'
+import { CtaLink, CtaAnchorLink } from '~components/common/call-to-action'
 import { FormatNumber } from '~components/utils/format'
 import overviewStyles from './overview.module.scss'
